Changing the way Bootstrap Sass package is ref'd
Aliasing "bootstrap-sass" to the github URL
git://github.com/twbs/bootstrap-sass.git is probably not a good idea.
This could create confusion with the bootstrap-sass package in
the official repository which actually points to
git://github.com/jlong/sass-twitter-bootstrap.git and only goes up to
version 3.0.2 currently. This is NOT the official package.

Changed references to "bootstrap-sass-official" to avoid any
potential confusion.