Make it easier to debug errors in .jsx files
It was very hard for me to understand what the error was from the current output ([#412](https://github.com/yeoman/generator-gulp-webapp/issues/412)).
Using console.log() will tell you what the error is, the file that triggers it and the exact line number. The gulp server will also not crash every time it outputs an error. Make sure you keep an eye on the console.